feat: adding support for floats, and negative numbers. Adding more verifications and errors. Cleaning the code

+++ cmd/operation/operations.go
@@ -1,18 +1,18 @@
 package operation
 
-func Add(a, b int) int {
+func Add(a, b float64) float64 {
 	return a + b
 }
 
-func Subtract(a, b int) int {
+func Subtract(a, b float64) float64 {
 	return a - b
 }
 
-func Multiply(a, b int) int {
+func Multiply(a, b float64) float64 {
 	return a * b
 }
 
-func Divide(a, b int) int {
+func Divide(a, b float64) float64 {
 	if b == 0 {
 		panic("division by zero")
 	}
